Added md5 hashing function

+++ Crypto/src/part.h
@@ -12,12 +12,12 @@
 
 #define BCAST_TAG 0
 #define CHARSET_SIZE 64
-#define DISPOSITIONS(cs_len, pass_len) (powl((cs_len),(pass_len)));
+#define DISPOSITIONS(cs_len, pass_len) (pow((cs_len),(pass_len)));
 #define DISP_PER_PROC(disp, num_procs) (((disp)/(num_procs))+1);
 #define HASH_SIZE 16
 #define PADDING 2
 #define STR_PASSLEN 3
-#define STARTING_CHAR(init, cs_size, pos) (((int)((init)/(powl(cs_size,pos))))%(cs_size))
+#define STARTING_CHAR(init, cs_size, pos) (((int)((init)/(pow(cs_size,pos))))%(cs_size))
 #define UI_FIELDS 3
 
 typedef struct string_t string_t;
